# openapi.yaml
openapi: 3.0.3
info:
  title: MySelfServe API
  version: 0.1.0
  description: |
    BC Government Income Assistance Self-Service Portal API.

    MySS 2.0 provides self-service capabilities for BC income assistance clients
    to manage service requests, monthly reports, payments, employment plans,
    messages, and account settings. Ministry workers can access client data
    through an admin support view.

    ## Authentication

    Most endpoints require a JWT Bearer token issued via BCeID/IDIR through
    Keycloak OIDC. The token is signed with HS256 and contains claims for
    `sub` (user_id), `role`, `bceid_guid`, and `idir_username`.

    ## Roles

    - **CLIENT** — BCeID-authenticated income assistance clients
    - **WORKER** — IDIR-authenticated ministry staff (includes ADMIN)
    - **ADMIN** — System administrators (inherits WORKER permissions)

    ## Error Handling

    Errors from the upstream Siebel/ICM system are mapped to standard HTTP
    status codes with structured JSON error bodies. A circuit breaker protects
    the application during ICM outages (5 failures → 30s open window → 503).
  contact:
    name: MySS Development Team
  license:
    name: BC Government
    url: https://www2.gov.bc.ca/
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://myss-api-dev.apps.silver.devops.gov.bc.ca
    description: Development (OpenShift)
  - url: https://myss-api-test.apps.silver.devops.gov.bc.ca
...
security:
  - BearerAuth: []
paths:
...
  /registration/{token}/step/{step_number}:
    get:
      operationId: getRegistrationStep
      tags: [Registration]
      summary: Get current registration step state
      security: []
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: step_number
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Step state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StepStateResponse"
...
  /account/profile:
    get:
      operationId: getProfile
      tags: [Account]
      summary: Get account profile
      responses:
        "200":
          description: Account profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountInfoResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /account/contact:
    patch:
      operationId: updateContact
      tags: [Account]
      summary: Update contact information
      description: Update email and/or phone numbers. Email requires confirmation field to match.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContactRequest"
      responses:
        "200":
          description: Contact updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /account/case-members:
    get:
      operationId: getCaseMembers
      tags: [Account]
      summary: Get case members
      responses:
        "200":
          description: Case members
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CaseMemberListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /account/post-login-sync:
    post:
      operationId: postLoginSync
      tags: [Account]
      summary: Trigger post-login data synchronization
      responses:
        "202":
          description: Sync accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: accepted
        "401":
          $ref: "#/components/responses/Unauthorized"

  # ──────────────────────────────────────────────
  # PIN Management
  # ──────────────────────────────────────────────
...
  /admin/support-view/search:
    post:
      operationId: supportViewSearch
      tags: ["Admin: Support View"]
      summary: Search for clients
      description: Workers can search by name or SIN to find clients.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SearchRequest"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                type: object
                description: Client search results
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/support-view/tombstone:
    post:
      operationId: createSupportViewSession
      tags: ["Admin: Support View"]
      summary: Create a support view session
      description: |
        Creates a 15-minute support view session for a client. Workers cannot
        create sessions for their own account (self-impersonation check).
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TombstoneRequest"
      responses:
        "200":
          description: Support view session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportViewSession"
        "400":
          description: Self-impersonation not allowed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    delete:
      operationId: endSupportViewSession
      tags: ["Admin: Support View"]
      summary: End a support view session
      parameters:
        - name: X-Support-View-Client
          in: header
          required: true
          schema:
            type: string
          description: Client BCeID GUID
      responses:
        "204":
          description: Session ended
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/support-view/client-data/{resource}:
    get:
      operationId: getSupportViewClientData
      tags: ["Admin: Support View"]
      summary: Get client data in support view
      description: Retrieves client data for the specified resource while in a support view session.
      parameters:
        - name: resource
          in: path
          required: true
          schema:
            type: string
          description: Resource type to retrieve (e.g. profile, service-requests, messages)
      responses:
        "200":
          description: Client data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SupportViewClientData"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ──────────────────────────────────────────────
  # Admin: AO Registration
  # ──────────────────────────────────────────────
  /admin/ao/login:
    post:
      operationId: aoLogin
      tags: ["Admin: AO Registration"]
      summary: Authenticate AO registration session
      description: Workers log in with a service request number and SIN to start an administrative override registration.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AOLoginRequest"
      responses:
        "200":
          description: AO session created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AOSessionToken"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: Invalid SR number or SIN
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /admin/ao/registration/{token}/step/{step}:
    get:
      operationId: getAORegistrationStep
      tags: ["Admin: AO Registration"]
      summary: Get AO registration step data
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: step
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Step data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AORegistrationStep"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

    post:
      operationId: submitAORegistrationStep
      tags: ["Admin: AO Registration"]
      summary: Submit AO registration step data
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
        - name: step
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Step submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AORegistrationStep"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/ao/form/{sr_id}/submit:
    post:
      operationId: submitAOForm
      tags: ["Admin: AO Registration"]
      summary: Submit an AO form
      parameters:
        - name: sr_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Form submitted
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /admin/ao/ia-applications:
    get:
      operationId: listAOApplications
      tags: ["Admin: AO Registration"]
      summary: List AO IA applications
      responses:
        "200":
          description: Applications list
          content:
            application/json:
              schema:
                type: object
                properties:
                  applications:
                    type: array
                    items:
                      type: object
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

# ════════════════════════════════════════════════
# Components
# ════════════════════════════════════════════════
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token signed with HS256. Claims: `sub` (user_id), `role` (CLIENT/WORKER/ADMIN),
        `bceid_guid` (clients), `idir_username` (workers/admins), `exp` (expiry).
    WebhookSecret:
      type: apiKey
      in: header
      name: X-Webhook-Secret
      description: HMAC secret for internal AV scanner webhook
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    ServiceUnavailable:
      description: Upstream service (ICM/Siebel) unavailable — circuit breaker open
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        detail:
          type: string
          description: Human-readable error message
      required:
        - detail
...
    AccountCreationType:
      type: string
      enum: [SELF, WITH_HELPER, POA, PARENT]
...
    PhoneNumberOperation:
      type: string
      enum: [ADD, UPDATE, DELETE]

    # ── Eligibility ─────────────────────────────
...
    StepStateResponse:
      type: object
      required:
        - token
        - step_reached
        - account_creation_type
      properties:
        token:
          type: string
        step_reached:
          type: integer
        account_creation_type:
          $ref: "#/components/schemas/AccountCreationType"
...
    AccountInfoResponse:
      type: object
      required:
        - user_id
      properties:
        user_id:
          type: string
        email:
          type: string
          nullable: true
        phone_numbers:
          type: array
          items:
            type: object
          default: []
        case_number:
          type: string
          nullable: true
        case_status:
          type: string
          nullable: true
    PhoneNumberUpdate:
      type: object
      required:
        - operation
      properties:
        phone_id:
          type: integer
          nullable: true
        phone_number:
          type: string
          default: ""
        phone_type:
          type: string
          default: ""
        operation:
          $ref: "#/components/schemas/PhoneNumberOperation"
    UpdateContactRequest:
      type: object
      properties:
        email:
          type: string
          nullable: true
        email_confirm:
          type: string
          nullable: true
          description: Must match email if both provided
        phones:
          type: array
          items:
            $ref: "#/components/schemas/PhoneNumberUpdate"
          default: []
    CaseMember:
      type: object
      required:
        - name
        - relationship
      properties:
        name:
          type: string
        relationship:
          type: string
    CaseMemberListResponse:
      type: object
      required:
        - members
      properties:
        members:
          type: array
          items:
            $ref: "#/components/schemas/CaseMember"

    # ── PIN ─────────────────────────────────────
...
    SearchRequest:
      type: object
      properties:
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        sin:
          type: string
          nullable: true
        page:
          type: integer
          default: 1
    TombstoneRequest:
      type: object
      required:
        - client_bceid_guid
      properties:
        client_bceid_guid:
          type: string
    SupportViewSession:
      type: object
      required:
        - session_token
        - client_bceid_guid
        - expires_at
      properties:
        session_token:
          type: string
          description: "Format: {worker_idir}:{client_bceid_guid}"
        client_bceid_guid:
          type: string
        expires_at:
          type: string
          format: date-time
    SupportViewClientData:
      type: object
      required:
        - resource
        - client_bceid_guid
        - client_portal_id
        - data
      properties:
        resource:
          type: string
        client_bceid_guid:
          type: string
        client_portal_id:
          type: string
        data:
          type: object
          additionalProperties: true

    # ── Admin: AO Registration ──────────────────
    AOLoginRequest:
      type: object
      required:
        - sr_number
        - sin
      properties:
        sr_number:
          type: string
        sin:
          type: string
    AOSessionToken:
      type: object
      required:
        - session_token
        - expires_at
      properties:
        session_token:
          type: string
        expires_at:
          type: string
          format: date-time
    AORegistrationStep:
      type: object
      required:
        - step
        - data
      properties:
        step:
          type: integer
        data:
          type: object
          additionalProperties: true